wip: refactor: restructure library
