+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './superface-test';
+export * from './hash';
+export * from './config';
+export * from './nock';
