+++ src/interfaces/matcher.ts
@@ -1,0 +1,8 @@
+import { RecordingDefinitions } from './nock';
+
+export interface IMatcher {
+  match: (
+    oldRecordings: RecordingDefinitions,
+    newRecordings: RecordingDefinitions
+  ) => Promise<boolean>;
+}
