+++ src/interfaces/nock.ts
@@ -1,0 +1,24 @@
+import { Definition } from 'nock/types';
+
+export interface NockConfig {
+  path?: string;
+  fixture?: string;
+  enableReqheadersRecording?: boolean;
+  testInstance?: unknown;
+}
+
+export type RecordingDefinition = Definition & {
+  rawHeaders?: string[];
+};
+export type RecordingDefinitions = RecordingDefinition[];
+
+export type ProcessingFunction = (
+  recordings: RecordingDefinitions
+) => Promise<void> | void;
+
+export interface RecordingProcessOptions {
+  processRecordings?: boolean;
+  beforeRecordingSave?: ProcessingFunction;
+  beforeRecordingLoad?: ProcessingFunction;
+  hideInput?: string[];
+}
