+++ src/interfaces/replace.ts
@@ -1,0 +1,7 @@
+import { RecordingDefinition } from './nock';
+
+export interface ReplaceOptions {
+  definition: RecordingDefinition;
+  credential: string;
+  placeholder: string;
+}
