+++ src/nock/matcher/matcher.ts
@@ -1,0 +1,8 @@
+import { RecordingDefinitions } from '../../interfaces';
+
+export function match(
+  _oldRecordings: RecordingDefinitions,
+  _newRecordings: RecordingDefinitions
+): boolean {
+  return true;
+}
