--- src/superface-test.interfaces.ts
@@ -1,66 +1,0 @@
-import {
-  Profile,
-  Provider,
-  Result,
-  SuperfaceClient,
-  UseCase,
-} from '@superfaceai/one-sdk';
-import {
-  NonPrimitive,
-  Primitive,
-} from '@superfaceai/one-sdk/dist/internal/interpreter/variables';
-import { Definition } from 'nock/types';
-
-export interface SuperfaceTestConfigPayload {
-  client?: SuperfaceClient;
-  profile?: Profile | string;
-  provider?: Provider | string;
-  useCase?: UseCase | string;
-}
-
-export type InputVariables = Record<string, Primitive>;
-
-export interface HashOptions {
-  input: NonPrimitive;
-  testName?: string;
-}
-
-export type SuperfaceTestRun = Omit<
-  SuperfaceTestConfigPayload,
-  'testInstance'
-> &
-  HashOptions;
-
-export interface SuperfaceTestConfig {
-  client?: SuperfaceClient;
-  profile?: Profile;
-  provider?: Provider;
-  useCase?: UseCase;
-}
-
-export type CompleteSuperfaceTestConfig = Required<SuperfaceTestConfig>;
-
-export type TestingReturn = Result<unknown, string>;
-
-export interface NockConfig {
-  path?: string;
-  fixture?: string;
-  enableReqheadersRecording?: boolean;
-  testInstance?: unknown;
-}
-
-export type RecordingDefinition = Definition & {
-  rawHeaders?: string[];
-};
-export type RecordingDefinitions = RecordingDefinition[];
-
-export type ProcessingFunction = (
-  recordings: RecordingDefinitions
-) => Promise<void> | void;
-
-export interface RecordingProcessOptions {
-  processRecordings?: boolean;
-  beforeRecordingSave?: ProcessingFunction;
-  beforeRecordingLoad?: ProcessingFunction;
-  hideInput?: string[];
-}
